version 0.1

+++ user_service_01/src/main/java/com/cagayake/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.cagayake.controller;
 
+import com.cagayake.feign.OrderFeign;
 import com.cagayake.mapper.UserMapper;
 import com.cagayake.pojo.User;
 import com.cagayake.service.UserService;
@@ -15,35 +16,54 @@ import java.util.List;
 
 
 @RestController
-@RequestMapping("/api/user")
+@RequestMapping()
 public class UserController {
 
     private UserService userService;
 
-    @RequestMapping(value = "/add/{username}/{password}/{sex}/{phone_number}/{email}",method = RequestMethod.POST)
-    public void addUser(@PathVariable String username,@PathVariable String password,@PathVariable String sex,@PathVariable String phone_number,@PathVariable String email){
+    private OrderFeign orderFeign;
+
+    @RequestMapping(value = "/add/{username}/{password}/{sex}/{phone_number}/{email}",method = RequestMethod.POST,produces = "application/json")
+    public Response addUser(@PathVariable String username,@PathVariable String password,@PathVariable String sex,@PathVariable String phone_number,@PathVariable String email){
         User user = new User(username,sex,password,phone_number,email);
         userService.addUser(user);
+        return new Response(200,"success",null);
     }
 
-    @RequestMapping(value = "/getAllUser",method = RequestMethod.GET)
-    public List<User> findALlUser(){
-       return userService.findAllUser();
+    @RequestMapping(value = "/getAllUser",method = RequestMethod.GET,produces = "application/json")
+    public Response findALlUser(){
+       return new Response(200,"success",userService.findAllUser());
     }
 
-    @RequestMapping(value = "/delete/{username}",method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String username){
-        userService.deleteUser(username);
+    @RequestMapping(value = "/delete/{username}",method = RequestMethod.DELETE,produces = "application/json")
+    public Response deleteUser(@PathVariable String username){
+        Response response = orderFeign.deleteOrderByUsername(username);
+        if (response.getCode()==200){
+            userService.deleteUser(username);
+        }
+        return new Response(200,"success",null);
     }
 
-    @RequestMapping(value = "/update/{username}/{password}/{sex}/{phone_number}/{email}")
-    public void updateUser(@PathVariable String username,@PathVariable String password,@PathVariable String sex,@PathVariable String phone_number,@PathVariable String email){
+    @RequestMapping(value = "/update/{username}/{password}/{sex}/{phone_number}/{email}",produces = "application/json")
+    public Response updateUser(@PathVariable String username,@PathVariable String password,@PathVariable String sex,@PathVariable String phone_number,@PathVariable String email){
         User user = new User(username,sex,password,phone_number,email);
         userService.updateUser(user);
+        return new Response(200,"success",null);
+    }
+
+    @RequestMapping(value = "/create/{username}")
+    public Response createOrder(@PathVariable String username){
+        Response order = orderFeign.createOrder(username);
+        return new Response(200,"The order is as follows ",orderFeign.findOrderByUsername(username).getData());
     }
 
     @Autowired
     public void setUserService(UserService userService) {
         this.userService = userService;
+    }
+
+    @Autowired
+    public void setOrderFeign(OrderFeign orderFeign) {
+        this.orderFeign = orderFeign;
     }
 }
